tests/Node-nodeName.html

<!DOCTYPE html>
<title>Node.nodeName</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="../web-dom-core#dom-element-nodename">
<script src="testharness.js"></script>
<div id="log"></div>
<script>
test(function() {
  // same as Element-tagName except nodeName ...
  // XXX what about non-Element nodes
  var HTMLNS = "http://www.w3.org/1999/xhtml",
      SVGNS = "http://www.w3.org/2000/svg"
  assert_equals(document.createElementNS(HTMLNS, "I").nodeName, "I")
  assert_equals(document.createElementNS(HTMLNS, "i").nodeName, "I")
  assert_equals(document.createElementNS(SVGNS, "svg").nodeName, "svg")
  assert_equals(document.createElementNS(SVGNS, "SVG").nodeName, "SVG")
})
</script>
